# Conflicts:
#	database/mil_oblik.db
#	src/main/java/ntukhpi/semit/militaryoblik/adapters/ReservistAdapter.java

+++ src/main/java/ntukhpi/semit/militaryoblik/javafxview/DocumentsAllController.java
@@ -4,72 +4,84 @@ import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
-import javafx.scene.control.Alert;
+import javafx.scene.control.Label;
 import javafx.scene.control.TableColumn;
 import javafx.scene.control.TableView;
 import javafx.scene.control.cell.PropertyValueFactory;
 import javafx.scene.text.Text;
 import ntukhpi.semit.militaryoblik.MilitaryOblikKhPIMain;
-import ntukhpi.semit.militaryoblik.adapters.DocumentsAdapter;
-import ntukhpi.semit.militaryoblik.adapters.EducationAdapter;
+import ntukhpi.semit.militaryoblik.adapters.DocumentAdapter;
 import ntukhpi.semit.militaryoblik.entity.Document;
 import ntukhpi.semit.militaryoblik.entity.fromasukhpi.Prepod;
+import ntukhpi.semit.militaryoblik.javafxutils.DataFormat;
 import ntukhpi.semit.militaryoblik.javafxutils.Popup;
 import ntukhpi.semit.militaryoblik.service.DocumentServiceImpl;
 import ntukhpi.semit.militaryoblik.service.PrepodServiceImpl;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
-import javax.print.Doc;
 import java.time.LocalDate;
-import java.util.ArrayList;
-import java.util.List;
 
 @Component
 public class DocumentsAllController {
 
+    // ========== Константи для переходу на форму з редагуванням/додаванням нових документів ==========
     private final static String DOCUMENTS_ADD_JAVAFX = "/javafxview/DocumentsEdit.fxml";
-    private final static String DOCUMENTS_EDIT_JAVAFX = "/javafxview/DocumentsEdit.fxml";
+    private final static String DOCUMENTS_EDIT_JAVAFX = DOCUMENTS_ADD_JAVAFX;
     private final static String DOCUMENTS_ADD_JAVAFX_TITLE = "Додати новий документ";
     private final static String DOCUMENTS_EDIT_JAVAFX_TITLE = "Редагувати документ";
 
+    // ========== Інтерактивні елементи форми ==========
+    @FXML
+    private TableColumn<DocumentAdapter, LocalDate> dateColumn;
 
     @FXML
-    private TableColumn<DocumentsAdapter, LocalDate> dateColumn;
+    private TableColumn<DocumentAdapter, String> givenColumn;
 
     @FXML
-    private TableColumn<DocumentsAdapter, String> givenColumn;
-
-    @FXML
-    private TableColumn<DocumentsAdapter, String> numberColumn;
+    private TableColumn<DocumentAdapter, String> numberColumn;
 
     @FXML
     private Text pibText;
 
     @FXML
-    private TableColumn<DocumentsAdapter, String> typeColumn;
+    private TableColumn<DocumentAdapter, String> typeColumn;
 
     @FXML
-    private TableView<DocumentsAdapter> docsTableView;
+    private TableView<DocumentAdapter> docsTableView;
 
-    private ObservableList<DocumentsAdapter> docsObservableList;
+    // Список для зображення всіх документів в таблиці
+    private ObservableList<DocumentAdapter> docsObservableList;
 
+    // Обраний в ReservistsAll викладач
     private Prepod selectedPrepod;
 
+    // ========== Сервіси для роботи з БД ==========
     @Autowired
     DocumentServiceImpl documentService;
 
     @Autowired
     PrepodServiceImpl prepodService;
 
-    private ObservableList<DocumentsAdapter> getDocumentsData() {
-        return FXCollections.observableArrayList(documentService.getAllDocumentByPrepod(selectedPrepod).stream().map(DocumentsAdapter::new).toList());
+
+    /**
+     * Отримує з БД список всіх документів конкретного викладача та перетворює його на ObservableList
+     * @return ObservableList з усіма документави викладача
+     */
+    private ObservableList<DocumentAdapter> getDocumentsData() {
+        return FXCollections.observableArrayList(documentService.getAllDocumentByPrepod(selectedPrepod).stream().map(DocumentAdapter::new).toList());
     }
 
+
+    /**
+     * Ініціалізація форми
+     */
     public void initialize() {
+        // Зберігання викладача на випадок зміни вибору в формі ReservistsAll
         selectedPrepod = prepodService.getPrepodById(ReservistsAllController.getSelectedPrepodId());
 
-        pibText.setText(MilitaryOblikKhPIMain.getPIB(selectedPrepod));
+        pibText.setText(DataFormat.getPIB(selectedPrepod));
+        docsTableView.setPlaceholder(new Label("Ця людина поки не має жодного документа"));
 
         typeColumn.setCellValueFactory(new PropertyValueFactory<>("type"));
         numberColumn.setCellValueFactory(new PropertyValueFactory<>("number"));
@@ -78,40 +90,86 @@ public class DocumentsAllController {
 
         docsObservableList = getDocumentsData();
 
-        updateTable(docsObservableList);
+        docsTableView.setItems(docsObservableList);
     }
 
-    private void updateTable(ObservableList<DocumentsAdapter> docs) {
-        docsTableView.setItems(docs);
+    /**
+     * Обновляє таблицю після доданих чи прибраних даних
+     */
+    public void refreshDocsTable() {
+        docsTableView.refresh();
     }
 
+
+    /**
+     * Додавання нового документу до таблиці та БД
+     * @param document Новий документ
+     */
+    public void addNewDocument(Document document) {
+        documentService.createDocument(document);
+
+        docsObservableList.add(new DocumentAdapter(document));
+        refreshDocsTable();
+    }
+
+
+    /**
+     * Обновлення документа в таблиці та БД
+     * @param oldDocument Старий зразок документа
+     * @param newDocument Новий зразок документа
+     */
+    public void updateDocument(DocumentAdapter oldDocument, Document newDocument) {
+        documentService.updateDocument(oldDocument.getId(), newDocument);
+
+        docsObservableList.remove(oldDocument);
+        docsObservableList.add(new DocumentAdapter(newDocument));
+        refreshDocsTable();
+    }
+
+    /**
+     * Видаляє обраний документ з таблиці та БД
+     * @param event
+     */
     @FXML
     void deleteSelectedRow(ActionEvent event) {
-        DocumentsAdapter selectedDocument = docsTableView.getSelectionModel().getSelectedItem();
+        DocumentAdapter selectedDocument = docsTableView.getSelectionModel().getSelectedItem();
 
         if (selectedDocument != null) {
-            docsObservableList.remove(selectedDocument);
-            documentService.deleteDocument(selectedDocument.getId());
+            if (Popup.deleteConfirmation()) {   // Питаємо користувача
+                docsObservableList.remove(selectedDocument);
+                documentService.deleteDocument(selectedDocument.getId());
+            }
         } else {
             Popup.noSelectedRowAlert();
         }
     }
 
+    /**
+     * Відкриває вікно додавання нового документа
+     * @param event
+     */
     @FXML
     void openAddWindow(ActionEvent event) {
         MilitaryOblikKhPIMain.openEditWindow(DOCUMENTS_ADD_JAVAFX, DOCUMENTS_ADD_JAVAFX_TITLE, this, null);
     }
 
+    /**
+     * Відкриває вікно зміни документа
+     * @param event
+     */
     @FXML
     void openEditWindow(ActionEvent event) {
-        DocumentsAdapter selectedDocument = docsTableView.getSelectionModel().getSelectedItem();
-        if (selectedDocument != null) {
+        DocumentAdapter selectedDocument = docsTableView.getSelectionModel().getSelectedItem();
+        if (selectedDocument != null)
             MilitaryOblikKhPIMain.openEditWindow(DOCUMENTS_EDIT_JAVAFX, DOCUMENTS_EDIT_JAVAFX_TITLE, this, selectedDocument);
-        } else {
+        else
             Popup.noSelectedRowAlert();
-        }
     }
 
+    /**
+     * Відкриває попереднє вікно
+     * @param event
+     */
     @FXML
     void returnToMainForm(ActionEvent event) {
         MilitaryOblikKhPIMain.showReservistsWindow();
